Overhaul config to use custom object
Previously, configs are handled by a dict with key values pairs
parsed directly from the TOML file. However, this structure requires
key-value pairs to actually exist which may not be the case
when Priviblur gets updated to a new version.

As such by swapping to an implementation via a custom object,
default keys can be defined (along with a default value) ensuring that
configuration fields will always exist