Add workaround for poisoned connection pool
https://github.com/syeopite/priviblur/issues/4 appears to be caused by
faulty connections that aren't properly released which will result in
a PoolTimeout exception when attempting to fetch one.

When enough of a pool gets poisoned, all requests will fail due to the
PoolTimeout exceptions.

This commit implements a workaround by detecting and then reinitializing
poisoned connection pools which would clear all of the faulty connections.