Use JavaScript (with fallback) to populate polls
In server-side rendering, as posts are rendered one by one sequentially
the results for polls must also be fetched sequentially.

This results in a lengthy load time as Priviblur has to send dozens of
requests one-by-one before the page can be returned.

By populating the polls client side, this lengthy fetch process can be
avoided entirely.

For JS-disabled users, a fetch_polls URL parameter is provided as a
fallback to allow for server-side fetching of poll results. However,
due to the aforementioned reasons above, this is only available on
individual posts.