Reduce number of requests when fetching blog post
When rendering a single blog post, two requests are made.
The first is used to fetch the post, while the second is used to fetch
the blog info and some additional posts from before the target post.

However, the first request already contains the blog info and Priviblur
currently does not use the "before posts". And as such, it can be
optimized down to a single request that will contain both the post and
blog info.