Refactor error handlers to use msg_error template
The current error handlers are set to process explicitly
defined and known errors which has set behaviors
for what to display to the end user.

It does not actually handle an unexpected error where it is
imperative to also display debug information.

As such the generic_error template as been renamed to msg_error
as to reflect how it is currently used within Priviblur