Add basic fallback error handler
This commit adds a proper basic fallback error handler that shows debeug
information and the context of when the error has occured.

This replaces the default Sanic error handler which does not show any debug
information when development mode is off.

The styling for the error handler is very basic and serves as a placeholder
until a proper design can be done for it.

This will likely be a v0.3.0 feature rather than v0.1.0 one.