Refactor naming scheme of parsers
BlogParser -> BlogTimelineParser
BlogInfoParser -> BlogParser

This new name is more accurate as to the actual parsed results
of the two parsers.

The old BlogInfoParser parses the actual blog that contains the name,
header, etc; while the BlogParser parses a simple object that stores
the blog info (which calls BlogInfoParser), a list of posts, and a cursor.