Overhaul routing logic to check slug before req.
This commit removes the slug check logic from the former route handlers
`_blog_post` and `_blog_post_no_slug`. Replacing this, is a middleware
that handles the slug check logic instead.

By using a middleware, we can apply this logic uniformly to all
additional blog post routes. For example for a post note viewer
ensuring that the endpoints to view post notes always has the correct
slug if applicable, and to remove the given slug if unnecessary.

As the logic that separated the slug and slugless post routes have
been transformed to a single middleware, these two routes have been
merged to use the same handler.