Fix reblog attribution linking to external domain
On blogs with a custom domain the url given in rebloggedFrom*
and rebloggedRoot* will use their custom domain instead of a
regular tumblr domain.

In order to preserve the slug of the post priviblur uses the url given
from rebloggedFrom* and rebloggedRoot* which meant that users can be
directed to an external site unknownly.

This commit fixes that by using the url given in the parentPostUrl
which should always be a regular tumblr link. Redundancy is also added
in that priviblur will check if the URL is a Tumblr domain and return
a span if it isn't and priviblur is unable to create a url to the original
post