CI: Generate locale data during testing
Generated locale data is not committed until a PR is merged.
This means that when newly added languages are enabled the test
instance ran by the CI attempts to fetch the data from an
unavailable language causing an error.

This PR generates the locale data prior to running the test instance.
This does mean that the locale data generation script is ran twice
once in the CI and once in the dedicated workflow but the script is
simple enough that this duplication is irrelevant and attempts to fix it
will complicate the workflows significantly